refactor: 대시보드 디렉토리 구조 리팩토링 (#63)
