+++ apps/web/src/app/(main)/dashboard/_components/UserHomeEmptyScreen.tsx
@@ -2,10 +2,10 @@
 
 import React from 'react';
 import { Flex } from '@repo/ui/Flex';
-import { UserHomeHeader } from '@web/app/(main)/_components/user/UserHomeHeader/UserHomeHeader';
-import { NoAnnouncements } from '@web/app/(main)/_components/user/EmptyState/NoAnnouncements';
-import { NoDocs } from '@web/app/(main)/_components/user/EmptyState/NoDocs';
-import { NoInterview } from '@web/app/(main)/_components/user/EmptyState/NoInterviews';
+import { UserHomeHeader } from '@web/app/(main)/dashboard/_components/user/UserHomeHeader/UserHomeHeader';
+import { NoAnnouncements } from '@web/app/(main)/dashboard/_components/user/EmptyState/NoAnnouncements';
+import { NoDocs } from '@web/app/(main)/dashboard/_components/user/EmptyState/NoDocs';
+import { NoInterview } from '@web/app/(main)/dashboard/_components/user/EmptyState/NoInterviews';
 
 export const UserHomeEmptyScreen: React.FC = () => {
   return (
