+++ apps/web/src/app/(main)/dashboard/_components/admin/OverallProgress/OverallProgress.css.ts
