+++ apps/web/src/app/(main)/dashboard/_components/user/EmptyState/EmptyState.css.ts
