+++ apps/web/src/app/(main)/dashboard/_components/user/UserAnnouncementProgress/UserAnnouncementProgress.tsx
